Moved RGBIO logic out of Com module and into its own.
